Moved the flight activity functionality into a fragment. Need to fix the sliding up panel ui bugs.
