+++ Android/src/org/droidplanner/android/notifications/StatusBarNotificationProvider.java
@@ -1,14 +1,5 @@
 package org.droidplanner.android.notifications;
 
-import org.beyene.sius.unit.length.LengthUnit;
-import org.beyene.sius.unit.length.Meter;
-import org.droidplanner.android.R;
-import org.droidplanner.android.DroidPlannerApp;
-import org.droidplanner.android.activities.FlightActivity;
-import org.droidplanner.android.utils.prefs.DroidPlannerPrefs;
-import org.droidplanner.android.utils.unit.UnitManager;
-import org.droidplanner.android.utils.unit.providers.length.LengthUnitProvider;
-
 import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
 import android.content.Context;
@@ -32,17 +23,24 @@ import com.o3dr.services.android.lib.drone.property.VehicleMode;
 import com.o3dr.services.android.lib.util.MathUtils;
 import com.o3dr.services.android.lib.util.SpannableUtils;
 
+import org.beyene.sius.unit.length.LengthUnit;
+import org.droidplanner.android.DroidPlannerApp;
+import org.droidplanner.android.R;
+import org.droidplanner.android.activities.FlightActivity;
+import org.droidplanner.android.utils.prefs.DroidPlannerPrefs;
+import org.droidplanner.android.utils.unit.UnitManager;
+
 /**
  * Implements DroidPlanner's status bar notifications.
  */
 public class StatusBarNotificationProvider implements NotificationHandler.NotificationProvider {
 
-	private static final String TAG = StatusBarNotificationProvider.class.getSimpleName();
+    private static final String TAG = StatusBarNotificationProvider.class.getSimpleName();
 
-	/**
-	 * Android status bar's notification id.
-	 */
-	private static final int NOTIFICATION_ID = 1;
+    /**
+     * Android status bar's notification id.
+     */
+    private static final int NOTIFICATION_ID = 1;
 
     /**
      * This is the period for the flight time update.
@@ -51,54 +49,55 @@ public class StatusBarNotificationProvider implements NotificationHandler.Notifi
 
     private final Handler mHandler = new Handler();
 
-	/**
-	 * Application context.
-	 */
-	private final Context mContext;
+    /**
+     * Application context.
+     */
+    private final Context mContext;
 
-	/**
-	 * Builder for the app notification.
-	 */
-	private NotificationCompat.Builder mNotificationBuilder;
+    /**
+     * Builder for the app notification.
+     */
+    private NotificationCompat.Builder mNotificationBuilder;
 
-	/**
-	 * Pending intent for the notification on click behavior. Opens the
-	 * FlightActivity screen.
-	 */
-	private final PendingIntent mNotificationIntent;
+    /**
+     * Pending intent for the notification on click behavior. Opens the
+     * FlightActivity screen.
+     */
+    private final PendingIntent mNotificationIntent;
 
-	/**
-	 * Pending intent for the notification connect/disconnect action.
-	 */
-	private final PendingIntent mToggleConnectionIntent;
+    /**
+     * Pending intent for the notification connect/disconnect action.
+     */
+    private final PendingIntent mToggleConnectionIntent;
 
-	/**
-	 * Uses to generate the inbox style use to populate the notification.
-	 */
-	private InboxStyleBuilder mInboxBuilder;
+    /**
+     * Uses to generate the inbox style use to populate the notification.
+     */
+    private InboxStyleBuilder mInboxBuilder;
 
-	/**
-	 * Handle to the app preferences.
-	 */
-	private final DroidPlannerPrefs mAppPrefs;
+    /**
+     * Handle to the app preferences.
+     */
+    private final DroidPlannerPrefs mAppPrefs;
 
     private final Drone drone;
 
-	StatusBarNotificationProvider(Context context, Drone api) {
-		mContext = context;
+    StatusBarNotificationProvider(Context context, Drone api) {
+        mContext = context;
         this.drone = api;
-		mAppPrefs = new DroidPlannerPrefs(context);
+        mAppPrefs = new DroidPlannerPrefs(context);
 
-		mNotificationIntent = PendingIntent.getActivity(mContext, 0, new Intent(mContext,
-				FlightActivity.class), 0);
+        mNotificationIntent = PendingIntent.getActivity(mContext, 0, new Intent(mContext,
+                FlightActivity.class), 0);
 
-		mToggleConnectionIntent = PendingIntent
+        mToggleConnectionIntent = PendingIntent
                 .getBroadcast(mContext, 0, new Intent(DroidPlannerApp.ACTION_TOGGLE_DRONE_CONNECTION), 0);
 
         LocalBroadcastManager.getInstance(context).registerReceiver(eventReceiver, eventFilter);
-	}
+    }
 
     private static final IntentFilter eventFilter = new IntentFilter();
+
     static {
         eventFilter.addAction(AttributeEvent.STATE_CONNECTED);
         eventFilter.addAction(AttributeEvent.BATTERY_UPDATED);
@@ -198,7 +197,7 @@ public class StatusBarNotificationProvider implements NotificationHandler.Notifi
             if (drone == null || !drone.isConnected())
                 return;
 
-            if(mInboxBuilder != null) {
+            if (mInboxBuilder != null) {
                 long timeInSeconds = drone.getFlightTime();
                 long minutes = timeInSeconds / 60;
                 long seconds = timeInSeconds % 60;
@@ -211,141 +210,141 @@ public class StatusBarNotificationProvider implements NotificationHandler.Notifi
         }
     };
 
-	private void updateRadio(Drone drone) {
-		if (mInboxBuilder == null)
-			return;
+    private void updateRadio(Drone drone) {
+        if (mInboxBuilder == null)
+            return;
 
         Signal droneSignal = drone.getAttribute(AttributeType.SIGNAL);
         String update = droneSignal == null ? "--" : String.format("%d%%", MathUtils.getSignalStrength(droneSignal
                 .getFadeMargin(), droneSignal.getRemFadeMargin()));
-		mInboxBuilder.setLine(4, SpannableUtils.normal("Signal:   ", SpannableUtils.bold(update)));
-	}
+        mInboxBuilder.setLine(4, SpannableUtils.normal("Signal:   ", SpannableUtils.bold(update)));
+    }
 
-	private void updateHome(Drone drone) {
-		if (mInboxBuilder == null)
-			return;
+    private void updateHome(Drone drone) {
+        if (mInboxBuilder == null)
+            return;
 
         String update = "--";
-            final Gps droneGps = this.drone.getAttribute(AttributeType.GPS);
-            final Home droneHome = this.drone.getAttribute(AttributeType.HOME);
-            if(droneGps != null && droneGps.isValid() && droneHome != null && droneHome.isValid()) {
-                LengthUnit distanceToHome = UnitManager.getUnitSystem(mContext).getLengthUnitProvider()
-                        .boxBaseValueToTarget(MathUtils.getDistance2D(droneHome.getCoordinate(), droneGps.getPosition()));
-                update = String.format("Home\n%s", distanceToHome);
-            }
-		mInboxBuilder.setLine(0, SpannableUtils.normal("Home:   ", update));
-	}
+        final Gps droneGps = this.drone.getAttribute(AttributeType.GPS);
+        final Home droneHome = this.drone.getAttribute(AttributeType.HOME);
+        if (droneGps != null && droneGps.isValid() && droneHome != null && droneHome.isValid()) {
+            LengthUnit distanceToHome = UnitManager.getUnitSystem(mContext).getLengthUnitProvider()
+                    .boxBaseValueToTarget(MathUtils.getDistance2D(droneHome.getCoordinate(), droneGps.getPosition()));
+            update = String.format("Home\n%s", distanceToHome);
+        }
+        mInboxBuilder.setLine(0, SpannableUtils.normal("Home:   ", update));
+    }
 
-	private void updateGps(Drone drone) {
-		if (mInboxBuilder == null)
-			return;
+    private void updateGps(Drone drone) {
+        if (mInboxBuilder == null)
+            return;
 
-      Gps droneGps = drone.getAttribute(AttributeType.GPS);
+        Gps droneGps = drone.getAttribute(AttributeType.GPS);
         String update = droneGps == null ? "--" : String.format(
                 "%d, %s", droneGps.getSatellitesCount(), droneGps.getFixType());
-		mInboxBuilder.setLine(1, SpannableUtils.normal("Satellite:   ", SpannableUtils.bold(update)));
-	}
+        mInboxBuilder.setLine(1, SpannableUtils.normal("Satellite:   ", SpannableUtils.bold(update)));
+    }
 
-	private void updateBattery(Drone drone) {
-		if (mInboxBuilder == null)
-			return;
+    private void updateBattery(Drone drone) {
+        if (mInboxBuilder == null)
+            return;
 
         Battery droneBattery = drone.getAttribute(AttributeType.BATTERY);
         String update = droneBattery == null ? "--" : String.format(
                 "%2.1fv (%2.0f%%)", droneBattery.getBatteryVoltage(),
                 droneBattery.getBatteryRemain());
 
-		mInboxBuilder.setLine(3, SpannableUtils.normal("Battery:   ", SpannableUtils.bold(update)));
-	}
+        mInboxBuilder.setLine(3, SpannableUtils.normal("Battery:   ", SpannableUtils.bold(update)));
+    }
 
-	private void updateDroneState(Drone drone) {
-		if (mInboxBuilder == null)
-			return;
+    private void updateDroneState(Drone drone) {
+        if (mInboxBuilder == null)
+            return;
 
         mHandler.removeCallbacks(mFlightTimeUpdater);
         if (drone != null && drone.isConnected()) {
             mFlightTimeUpdater.run();
         }
-	}
+    }
 
-	private void updateFlightMode(Drone drone) {
-		if (mNotificationBuilder == null)
-			return;
+    private void updateFlightMode(Drone drone) {
+        if (mNotificationBuilder == null)
+            return;
 
         State droneState = drone.getAttribute(AttributeType.STATE);
         VehicleMode mode = droneState == null ? null : droneState.getVehicleMode();
         String update = mode == null ? "--" : mode.getLabel();
-		final CharSequence modeSummary = SpannableUtils.normal("Flight Mode:  ", SpannableUtils.bold(update));
-		mNotificationBuilder.setContentTitle(modeSummary);
-	}
+        final CharSequence modeSummary = SpannableUtils.normal("Flight Mode:  ", SpannableUtils.bold(update));
+        mNotificationBuilder.setContentTitle(modeSummary);
+    }
 
-	/**
-	 * Build a notification from the notification builder, and display it.
-	 */
-	private void showNotification() {
-		if (mNotificationBuilder == null) {
-			return;
-		}
+    /**
+     * Build a notification from the notification builder, and display it.
+     */
+    private void showNotification() {
+        if (mNotificationBuilder == null) {
+            return;
+        }
 
-		if (mInboxBuilder != null) {
-			mNotificationBuilder.setStyle(mInboxBuilder.generateInboxStyle());
-		}
+        if (mInboxBuilder != null) {
+            mNotificationBuilder.setStyle(mInboxBuilder.generateInboxStyle());
+        }
 
-		NotificationManagerCompat.from(mContext).notify(NOTIFICATION_ID,
+        NotificationManagerCompat.from(mContext).notify(NOTIFICATION_ID,
                 mNotificationBuilder.build());
-	}
+    }
 
-	/**
-	 * Dismiss the app status bar notification.
-	 */
+    /**
+     * Dismiss the app status bar notification.
+     */
     @Override
-	public void onTerminate() {
+    public void onTerminate() {
         LocalBroadcastManager.getInstance(mContext).unregisterReceiver(eventReceiver);
-		NotificationManagerCompat.from(mContext).cancelAll();
-	}
+        NotificationManagerCompat.from(mContext).cancelAll();
+    }
 
-	private static class InboxStyleBuilder {
-		private static final int MAX_LINES_COUNT = 5;
+    private static class InboxStyleBuilder {
+        private static final int MAX_LINES_COUNT = 5;
 
-		private final CharSequence[] mLines = new CharSequence[MAX_LINES_COUNT];
+        private final CharSequence[] mLines = new CharSequence[MAX_LINES_COUNT];
 
-		private CharSequence mSummary;
+        private CharSequence mSummary;
 
-		private boolean mHasContent = false;
+        private boolean mHasContent = false;
 
-		public void setLine(int index, CharSequence content) {
-			if (index >= mLines.length || index < 0) {
-				Log.w(TAG, "Invalid index (" + index + ") for inbox content.");
-				return;
-			}
+        public void setLine(int index, CharSequence content) {
+            if (index >= mLines.length || index < 0) {
+                Log.w(TAG, "Invalid index (" + index + ") for inbox content.");
+                return;
+            }
 
-			mLines[index] = content;
-			mHasContent = true;
-		}
+            mLines[index] = content;
+            mHasContent = true;
+        }
 
-		public InboxStyleBuilder setSummary(CharSequence summary) {
-			mSummary = summary;
-			mHasContent = true;
-			return this;
-		}
+        public InboxStyleBuilder setSummary(CharSequence summary) {
+            mSummary = summary;
+            mHasContent = true;
+            return this;
+        }
 
-		public NotificationCompat.InboxStyle generateInboxStyle() {
-			if (!mHasContent) {
-				return null;
-			}
+        public NotificationCompat.InboxStyle generateInboxStyle() {
+            if (!mHasContent) {
+                return null;
+            }
 
-			NotificationCompat.InboxStyle inboxStyle = new NotificationCompat.InboxStyle();
-			if (mSummary != null) {
-				inboxStyle.setSummaryText(mSummary);
-			}
+            NotificationCompat.InboxStyle inboxStyle = new NotificationCompat.InboxStyle();
+            if (mSummary != null) {
+                inboxStyle.setSummaryText(mSummary);
+            }
 
-			for (CharSequence line : mLines) {
-				if (line != null) {
-					inboxStyle.addLine(line);
-				}
-			}
+            for (CharSequence line : mLines) {
+                if (line != null) {
+                    inboxStyle.addLine(line);
+                }
+            }
 
-			return inboxStyle;
-		}
-	}
+            return inboxStyle;
+        }
+    }
 }
